weather/requirements.txt
python-decouple
influxdb
influxdb-client